structured the all services in proper manner
