+++ Services/Multer/config.js
@@ -1,0 +1,12 @@
+import multer from "multer";
+
+/* FILE STORAGE */
+const storage = multer.diskStorage({
+  destination: (req, file, cb) => {
+    cb(null, "./public/images");
+  },
+  filename: (req, file, cb) => {
+    cb(null, file.originalname);
+  },
+});
+export const upload = multer({ storage });
